backend/requirements.txt
fastapi
uvicorn
google-cloud-spanner
google-cloud-aiplatform
google-adk
pydantic
pydantic-settings
python-dotenv
httpx
websockets
